# Conflicts:
#	services/self-service/src/main/java/com/epam/dlab/backendapi/SelfServiceApplicationConfiguration.java
#	services/self-service/src/main/java/com/epam/dlab/backendapi/core/UserInstanceDTO.java
#	services/self-service/src/main/java/com/epam/dlab/backendapi/resources/dto/ExploratoryCreateFormDTO.java

+++ infrastructure-provisioning/aws/src/base/dlab/aws_actions.py
@@ -23,15 +23,10 @@ import sys
 import os
 import json
 from fabric.api import *
+from fabric.contrib.files import exists
 import logging
 from dlab.aws_meta import *
 import traceback
-
-local_log_filename = "%s.log" % os.environ['request_id']
-local_log_filepath = "/response/" + local_log_filename
-logging.basicConfig(format='%(levelname)-8s [%(asctime)s]  %(message)s',
-                    level=logging.DEBUG,
-                    filename=local_log_filepath)
 
 
 def put_to_bucket(bucket_name, local_file, destination_file):
@@ -177,6 +172,8 @@ def create_iam_role(role_name, role_profile):
     try:
         conn.create_role(RoleName=role_name, AssumeRolePolicyDocument='{"Version":"2012-10-17","Statement":[{"Effect":"Allow","Principal":{"Service":["ec2.amazonaws.com"]},"Action":["sts:AssumeRole"]}]}')
         conn.create_instance_profile(InstanceProfileName=role_profile)
+        waiter = conn.get_waiter('instance_profile_exists')
+        waiter.wait(InstanceProfileName=role_profile)
     except botocore.exceptions.ClientError as e_role:
         if e_role.response['Error']['Code'] == 'EntityAlreadyExists':
             print "Instance profile already exists. Reusing..."
@@ -244,7 +241,7 @@ def remove_ec2(tag_name, tag_value):
                 waiter.wait(InstanceIds=[instance.id])
                 print "The instance " + instance.id + " has been terminated successfully"
         else:
-            print "There are no instances with " + tag_value + " name to terminate"
+            print "There are no instances with '" + tag_name + "' tag to terminate"
     except Exception as err:
         logging.info("Unable to remove EC2: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -344,43 +341,74 @@ def remove_roles_and_profiles(role_name, role_profile_name):
 
 def remove_all_iam_resources(instance_type, scientist=''):
     try:
-        print "[Removing IAM roles, instance profiles and policies]"
         client = boto3.client('iam')
         roles_list = []
-        for item in client.list_roles().get("Roles"):
+        for item in client.list_roles(MaxItems=250).get("Roles"):
             if os.environ['conf_service_base_name'] in item.get("RoleName"):
                 roles_list.append(item.get('RoleName'))
         if roles_list:
+            roles_list.sort(reverse=True)
             for iam_role in roles_list:
                 if '-ssn-Role' in iam_role:
                     if instance_type == 'ssn' or instance_type == 'all':
+                        try:
+                            client.delete_role_policy(RoleName=iam_role, PolicyName=os.environ['conf_service_base_name'] + '-ssn-Policy')
+                        except:
+                            print 'There is no policy ' + os.environ['conf_service_base_name'] + '-ssn-Policy to delete'
                         role_profile_name = os.environ['conf_service_base_name'] + '-ssn-Profile'
-                        client.delete_role_policy(RoleName=iam_role, PolicyName=os.environ['conf_service_base_name'] + '-ssn-Policy')
-                        remove_roles_and_profiles(iam_role, role_profile_name)
+                        try:
+                            client.get_instance_profile(InstanceProfileName=role_profile_name)
+                            remove_roles_and_profiles(iam_role, role_profile_name)
+                        except:
+                            print "There is no instance profile for " + iam_role
+                            client.delete_role(RoleName=iam_role)
+                            print "The IAM role " + iam_role + " has been deleted successfully"
                 if '-edge-Role' in iam_role:
                     if instance_type == 'edge' and scientist in iam_role:
                         remove_detach_iam_policies(iam_role, 'delete')
                         role_profile_name = os.environ['conf_service_base_name'] + '-' + '{}'.format(scientist) + '-edge-Profile'
-                        remove_roles_and_profiles(iam_role, role_profile_name)
+                        try:
+                            client.get_instance_profile(InstanceProfileName=role_profile_name)
+                            remove_roles_and_profiles(iam_role, role_profile_name)
+                        except:
+                            print "There is no instance profile for " + iam_role
+                            client.delete_role(RoleName=iam_role)
+                            print "The IAM role " + iam_role + " has been deleted successfully"
                     if instance_type == 'all':
                         remove_detach_iam_policies(iam_role, 'delete')
                         role_profile_name = client.list_instance_profiles_for_role(RoleName=iam_role).get('InstanceProfiles')
-                        for i in role_profile_name:
-                            role_profile_name = i.get('InstanceProfileName')
-                            remove_roles_and_profiles(iam_role, role_profile_name)
+                        if role_profile_name:
+                            for i in role_profile_name:
+                                role_profile_name = i.get('InstanceProfileName')
+                                remove_roles_and_profiles(iam_role, role_profile_name)
+                        else:
+                            print "There is no instance profile for " + iam_role
+                            client.delete_role(RoleName=iam_role)
+                            print "The IAM role " + iam_role + " has been deleted successfully"
                 if '-nb-Role' in iam_role:
                     if instance_type == 'notebook' and scientist in iam_role:
                         remove_detach_iam_policies(iam_role)
                         role_profile_name = os.environ['conf_service_base_name'] + '-' + "{}".format(scientist) + '-nb-Profile'
-                        remove_roles_and_profiles(iam_role, role_profile_name)
+                        try:
+                            client.get_instance_profile(InstanceProfileName=role_profile_name)
+                            remove_roles_and_profiles(iam_role, role_profile_name)
+                        except:
+                            print "There is no instance profile for " + iam_role
+                            client.delete_role(RoleName=iam_role)
+                            print "The IAM role " + iam_role + " has been deleted successfully"
                     if instance_type == 'all':
                         remove_detach_iam_policies(iam_role)
                         role_profile_name = client.list_instance_profiles_for_role(RoleName=iam_role).get('InstanceProfiles')
-                        for i in role_profile_name:
-                            role_profile_name = i.get('InstanceProfileName')
-                            remove_roles_and_profiles(iam_role, role_profile_name)
+                        if role_profile_name:
+                            for i in role_profile_name:
+                                role_profile_name = i.get('InstanceProfileName')
+                                remove_roles_and_profiles(iam_role, role_profile_name)
+                        else:
+                            print "There is no instance profile for " + iam_role
+                            client.delete_role(RoleName=iam_role)
+                            print "The IAM role " + iam_role + " has been deleted successfully"
         else:
-            print "There is no IAM role to delete"
+            print "There are no IAM roles instance profiles and policies to delete"
     except Exception as err:
         logging.info("Unable to remove some of the IAM resources: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -391,8 +419,14 @@ def remove_all_iam_resources(instance_type, scientist=''):
 
 
 def s3_cleanup(bucket, cluster_name, user_name):
+    s3_res = boto3.resource('s3')
+    client = boto3.client('s3')
     try:
-        s3_res = boto3.resource('s3')
+        client.head_bucket(Bucket=bucket)
+    except:
+        print "There is no bucket " + bucket + " or you do not permission to access it"
+        sys.exit(0)
+    try:
         resource = s3_res.Bucket(bucket)
         prefix = user_name + '/' + cluster_name + "/"
         for i in resource.objects.filter(Prefix=prefix):
@@ -408,7 +442,6 @@ def s3_cleanup(bucket, cluster_name, user_name):
 
 def remove_s3(bucket_type='all', scientist=''):
     try:
-        print "[Removing S3 buckets]"
         client = boto3.client('s3')
         bucket_list = []
         if bucket_type == 'ssn':
@@ -426,7 +459,6 @@ def remove_s3(bucket_type='all', scientist=''):
             if list_obj is not None:
                 for o in list_obj:
                     list_obj = o.get('Key')
-                    print list_obj
                     client.delete_objects(
                         Bucket=s3bucket,
                         Delete={'Objects': [{'Key': list_obj}]}
@@ -446,16 +478,17 @@ def remove_s3(bucket_type='all', scientist=''):
 
 def remove_subnets(tag_value):
     try:
-        print "[Removing subnets]"
         ec2 = boto3.resource('ec2')
         client = boto3.client('ec2')
         tag_name = os.environ['conf_service_base_name'] + '-Tag'
         subnets = ec2.subnets.filter(
             Filters=[{'Name': 'tag:{}'.format(tag_name), 'Values': [tag_value]}])
-        for subnet in subnets:
-            print subnet.id
-            client.delete_subnet(SubnetId=subnet.id)
-            print "The subnet " + subnet.id + " has been deleted successfully"
+        if subnets:
+            for subnet in subnets:
+                client.delete_subnet(SubnetId=subnet.id)
+                print "The subnet " + subnet.id + " has been deleted successfully"
+        else:
+            print "There are no private subnets to delete"
     except Exception as err:
         logging.info("Unable to remove subnet: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -467,16 +500,17 @@ def remove_subnets(tag_value):
 
 def remove_sgroups(tag_value):
     try:
-        print "[Removing security groups]"
         ec2 = boto3.resource('ec2')
         client = boto3.client('ec2')
         tag_name = os.environ['conf_service_base_name']
         sgs = ec2.security_groups.filter(
             Filters=[{'Name': 'tag:{}'.format(tag_name), 'Values': [tag_value]}])
-        for sg in sgs:
-            print sg.id
-            client.delete_security_group(GroupId=sg.id)
-            print "The security group " + sg.id + " has been deleted successfully"
+        if sgs:
+            for sg in sgs:
+                client.delete_security_group(GroupId=sg.id)
+                print "The security group " + sg.id + " has been deleted successfully"
+        else:
+            print "There are no security groups to delete"
     except Exception as err:
         logging.info("Unable to remove SG: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -488,14 +522,17 @@ def remove_sgroups(tag_value):
 
 def deregister_image(scientist):
     try:
-        print "[De-registering images]"
         client = boto3.client('ec2')
         response = client.describe_images(
-            Filters=[{'Name': 'name', 'Values': ['{}-{}-notebook-image'.format(os.environ['conf_service_base_name'], scientist)]}])
+            Filters=[{'Name': 'name', 'Values': ['{}-{}-*'.format(os.environ['conf_service_base_name'], scientist)]},
+                     {'Name': 'tag-value', 'Values': [os.environ['conf_service_base_name']]}])
         images_list = response.get('Images')
-        for i in images_list:
-            client.deregister_image(ImageId=i.get('ImageId'))
-            print "Notebook AMI " + i + " has been deregistered successfully"
+        if images_list:
+            for i in images_list:
+                client.deregister_image(ImageId=i.get('ImageId'))
+                print "Notebook AMI " + i.get('ImageId') + " has been deregistered successfully"
+        else:
+            print "There is no notebook ami to deregister"
     except Exception as err:
         logging.info("Unable to de-register image: " + str(err) + "\n Traceback: " + traceback.print_exc(file=sys.stdout))
         with open("/root/result.json", 'w') as result:
@@ -538,6 +575,9 @@ def remove_kernels(emr_name, tag_name, nb_tag_value, ssh_user, key_path, emr_ver
                 env.host_string = env.user + "@" + env.hosts
                 sudo('rm -rf  /opt/' + emr_version + '/' + emr_name + '/')
                 sudo('rm -rf /home/{}/.local/share/jupyter/kernels/*_{}'.format(ssh_user, emr_name))
+                if exists('/home/ubuntu/.ensure_dir/rstudio_emr_ensured'):
+                    sudo("sed -i '/" + emr_name + "/d' /home/ubuntu/.Renviron")
+                    sudo("sed -i 's|/opt/" + emr_version + '/' + emr_name + "/spark//R/lib:||g' /home/ubuntu/.bashrc")
                 print "Notebook's " + env.hosts + " kernels were removed"
         else:
             print "There are no notebooks to clean kernels."
@@ -558,9 +598,9 @@ def remove_route_tables(tag_name):
             if rtable:
                 rtable = rtable.get('RouteTableId')
                 client.delete_route_table(RouteTableId=rtable)
-                print "Route table " + rtable + " was removed"
+                print "Route table " + rtable + " has been removed"
             else:
-                print "There is no route table to remove"
+                print "There are no route tables to remove"
     except Exception as err:
         logging.info("Unable to remove route table: " + str(err) + "\n Traceback: " + traceback.print_exc(
             file=sys.stdout))
